base/requirements.txt
beautifulsoup4==4.12.2
black @ git+https://github.com/psf/black@e974fc3c52959e9f01bf62bdcd0d8c100ce78985
blurhash==1.1.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cups==0.0.6
decorator==5.1.1
gphoto2==2.5.0
idna==3.4
Mastodon.py==1.8.1
mypy-extensions==1.0.0
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
python-dateutil==2.8.2
python-magic==0.4.27
requests==2.31.0
RPi.GPIO==0.7.1
six==1.16.0
soupsieve==2.5
urllib3==2.0.4
